// backend/src/models/FormParserItem.ts
export interface FormParserItem {
  userId: string
  id: string
  createdAt: string
  ClientRef: string
  AGE_AT_TEST: string
  AVERAGE_STRESS: string
  CUBE_DIMENSION: string
  DATE_CAST: string
  DATE_TESTED: string
  GRADE: string
  LOCATION: string
  MAX_LOAD: string
  MODE_OF_FAILURE: string
  SPECIFIED_STRENGTH: string
  STRESS_FALURE: string
  SUPPLIER: string
  S3FILE_ID?: string
}